refactor(data): split landing-pages.ts into per-page modules (#544)
Split the 2737-line src/lib/data/landing-pages.ts into a
src/lib/data/landing-pages/ directory: one module per landing-page slug
(all four locales each), a types module, and an index.ts entry point that
re-exports the unchanged public API and carries the lookup map, slug list,
and helpers (getLandingPage/getLandingPageOrThrow/getAllLandingPages).

Pure mechanical extraction, zero behavior change: all 24 content objects
moved byte-for-byte (verified per-const), the public surface is re-exported
identically, and all 26 consumers import via the unchanged
$lib/data/landing-pages specifier (SvelteKit resolves it to index.ts).

Note: scripts/check-file-length.sh excludes src/lib/data/, so this file was
not actually flagged by `make file-length`; the split is a code-health win
that does not move that gate (see task report).
