+++ src/lib/data/landing-pages/community-first-event-platform.ts
@@ -1,0 +1,449 @@
+import type { LandingPageContent } from './types';
+
+export const communityFirstEventPlatformEN: LandingPageContent = {
+	slug: 'community-first-event-platform',
+	locale: 'en',
+	meta: {
+		title: 'Community-First Event Platform – Beyond Ticketing | Revel',
+		description:
+			'Build lasting communities, not just events. Organizations, membership tiers, potluck coordination, and more. Self-host for free or use our hosted version.',
+		keywords:
+			'community event platform, membership management, organization events, potluck coordination, community building, meetup alternative'
+	},
+	hero: {
+		headline: 'Build Communities, Not Just Events',
+		subheadline:
+			'Organizations, memberships, and unique tools like potluck coordination. Revel helps you foster lasting community connections.'
+	},
+	intro: {
+		paragraphs: [
+			'Most event platforms treat every gathering as a one-off transaction. But real communities need more than that—they need structure, continuity, and tools that support ongoing relationships.',
+			'Revel is built for communities first. Create organizations with membership tiers, assign roles and permissions, and use features like our unique potluck coordination system to make events more collaborative and less work for organizers.',
+			"Whether you're running a book club, hobby group, professional network, or maker space, Revel gives you the infrastructure to grow from casual meetups into a thriving community—without expensive subscriptions or platform lock-in."
+		]
+	},
+	features: [
+		{
+			icon: 'users',
+			title: 'Organization Structure',
+			description:
+				'Create community hubs with membership tiers, not just event listings. Owner, staff, and member roles with granular permissions.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Potluck Coordination',
+			description:
+				'Built-in system for coordinating who brings what. Handle dietary restrictions, quantities, and item management—no more messy spreadsheets.'
+		},
+		{
+			icon: 'shield',
+			title: 'Member-Only Events',
+			description:
+				'Restrict events to members, specific tiers, or keep them public. Build exclusive spaces for your community.'
+		},
+		{
+			icon: 'ticket',
+			title: 'Integrated Ticketing',
+			description:
+				'Free events, paid tickets, RSVP-only, or hybrid. Handle everything from casual meetups to professional conferences.'
+		},
+		{
+			icon: 'eye',
+			title: 'No Ads, Ever',
+			description:
+				'Your community deserves better than being shown ads. Self-host for free or use our ad-free hosted version.'
+		},
+		{
+			icon: 'code',
+			title: 'Open Source (MIT)',
+			description:
+				'Free to use, modify, and deploy. No vendor lock-in. Run on your own servers with complete control.'
+		}
+	],
+	benefits: {
+		title: 'Why Community Organizers Choose Revel',
+		items: [
+			'Build lasting membership structures, not just event-by-event lists',
+			'Coordinate potlucks and shared responsibilities effortlessly',
+			'No expensive subscriptions like Meetup ($540/year)',
+			'Ad-free experience for your members',
+			'Member-only events for building exclusive communities',
+			'Self-host for complete control and zero platform fees'
+		]
+	},
+	cta: {
+		title: 'Ready to Build Your Community?',
+		description: 'See Revel in action or deploy it yourself. No credit card required.',
+		buttons: [
+			{ text: 'Try the Live Demo', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{ text: 'Self-Host (GitHub)', href: 'https://github.com/letsrevel', variant: 'secondary' },
+			{ text: 'Contact Us', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'How is this different from Meetup?',
+			answer:
+				'Meetup charges $540/year and shows ads to your members. Revel is free to self-host and ad-free on our hosted version. We also integrate full ticketing capabilities and unique features like potluck coordination that Meetup lacks.'
+		},
+		{
+			question: 'What is potluck coordination?',
+			answer:
+				"It's a built-in system that lets attendees coordinate who brings what to events. Handle dietary restrictions, quantity management, and item assignments—no more messy spreadsheets or external tools. It's great for potlucks, equipment sharing, volunteer coordination, and more."
+		},
+		{
+			question: 'Can I create member-only events?',
+			answer:
+				'Yes. You can create organizations with membership tiers and restrict events to members only, specific membership levels, or keep them public. You have complete control over visibility and access.'
+		},
+		{
+			question: 'Is Revel really free?',
+			answer:
+				'Yes for self-hosting (MIT licensed). Our hosted version charges a small fee only for paid ticket sales (1.5% + €0.25 per ticket). Free events and RSVP-only events have zero platform fees on either version.'
+		},
+		{
+			question: 'What types of communities use Revel?',
+			answer:
+				"Book clubs, running groups, maker spaces, professional networks, hobby communities, special interest groups, and more. Any community that wants more than just basic event listings benefits from Revel's organization and membership features."
+		}
+	],
+	relatedPages: ['eventbrite-alternative', 'self-hosted-event-platform']
+};
+
+export const communityFirstEventPlatformDE: LandingPageContent = {
+	slug: 'community-first-event-platform',
+	locale: 'de',
+	meta: {
+		title: 'Community-First Event-Plattform – Mehr als Ticketing | Revel',
+		description:
+			'Baue dauerhafte Communities, nicht nur Events. Organisationen, Mitgliedschaftsstufen, Potluck-Koordination und mehr. Selbst hosten oder gehostete Version nutzen.',
+		keywords:
+			'community event plattform, mitgliederverwaltung, organisation events, potluck koordination, community aufbau, meetup alternative'
+	},
+	hero: {
+		headline: 'Baue Communities, Nicht Nur Events',
+		subheadline:
+			'Organisationen, Mitgliedschaften und einzigartige Tools wie Potluck-Koordination. Revel hilft dir, dauerhafte Community-Verbindungen aufzubauen.'
+	},
+	intro: {
+		paragraphs: [
+			'Die meisten Event-Plattformen behandeln jedes Treffen als einmalige Transaktion. Aber echte Communities brauchen mehr—sie brauchen Struktur, Kontinuität und Tools, die fortlaufende Beziehungen unterstützen.',
+			'Revel ist zuerst für Communities gebaut. Erstelle Organisationen mit Mitgliedschaftsstufen, weise Rollen und Berechtigungen zu und nutze Features wie unser einzigartiges Potluck-Koordinationssystem, um Events kollaborativer zu machen und weniger Arbeit für Organisatoren.',
+			'Egal ob du einen Buchclub, eine Hobbygruppe, ein professionelles Netzwerk oder einen Makerspace leitest, Revel gibt dir die Infrastruktur, um von lockeren Meetups zu einer florierenden Community zu wachsen—ohne teure Abos oder Plattform-Lock-in.'
+		]
+	},
+	features: [
+		{
+			icon: 'users',
+			title: 'Organisationsstruktur',
+			description:
+				'Erstelle Community-Hubs mit Mitgliedschaftsstufen, nicht nur Event-Listings. Owner-, Staff- und Member-Rollen mit granularen Berechtigungen.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Potluck-Koordination',
+			description:
+				'Integriertes System zur Koordination, wer was mitbringt. Verwalte Ernährungseinschränkungen, Mengen und Items—keine chaotischen Spreadsheets mehr.'
+		},
+		{
+			icon: 'shield',
+			title: 'Nur-Mitglieder Events',
+			description:
+				'Beschränke Events auf Mitglieder, spezifische Stufen oder halte sie öffentlich. Baue exklusive Räume für deine Community.'
+		},
+		{
+			icon: 'ticket',
+			title: 'Integriertes Ticketing',
+			description:
+				'Gratis-Events, bezahlte Tickets, nur RSVP oder hybrid. Verwalte alles von lockeren Meetups bis zu professionellen Konferenzen.'
+		},
+		{
+			icon: 'eye',
+			title: 'Niemals Werbung',
+			description:
+				'Deine Community verdient besser als Werbung zu sehen. Selbst hosten gratis oder nutze unsere werbefreie gehostete Version.'
+		},
+		{
+			icon: 'code',
+			title: 'Open Source (MIT)',
+			description:
+				'Kostenlos zu nutzen, modifizieren und deployen. Kein Vendor Lock-in. Betreibe auf eigenen Servern mit vollständiger Kontrolle.'
+		}
+	],
+	benefits: {
+		title: 'Warum Community-Organisatoren Revel Wählen',
+		items: [
+			'Baue dauerhafte Mitgliedschaftsstrukturen, nicht nur Event-Listen',
+			'Koordiniere Potlucks und geteilte Verantwortlichkeiten mühelos',
+			'Keine teuren Abos wie Meetup (540€/Jahr)',
+			'Werbefreie Erfahrung für deine Mitglieder',
+			'Nur-Mitglieder Events zum Aufbau exklusiver Communities',
+			'Selbst hosten für komplette Kontrolle und null Plattformgebühren'
+		]
+	},
+	cta: {
+		title: 'Bereit, Deine Community Aufzubauen?',
+		description: 'Sieh Revel in Aktion oder deploye es selbst. Keine Kreditkarte erforderlich.',
+		buttons: [
+			{ text: 'Live Demo Testen', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{
+				text: 'Selbst Hosten (GitHub)',
+				href: 'https://github.com/letsrevel',
+				variant: 'secondary'
+			},
+			{ text: 'Kontakt', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'Wie unterscheidet sich das von Meetup?',
+			answer:
+				'Meetup kostet 540€/Jahr und zeigt deinen Mitgliedern Werbung. Revel ist kostenlos zum Selbst-Hosten und werbefrei in unserer gehosteten Version. Wir integrieren auch vollständige Ticketing-Funktionen und einzigartige Features wie Potluck-Koordination, die Meetup fehlen.'
+		},
+		{
+			question: 'Was ist Potluck-Koordination?',
+			answer:
+				'Es ist ein integriertes System, das Teilnehmern ermöglicht zu koordinieren, wer was zu Events mitbringt. Verwalte Ernährungseinschränkungen, Mengen-Management und Item-Zuweisungen—keine chaotischen Spreadsheets oder externe Tools mehr. Es ist großartig für Potlucks, Equipment-Sharing, Freiwilligen-Koordination und mehr.'
+		},
+		{
+			question: 'Kann ich Nur-Mitglieder Events erstellen?',
+			answer:
+				'Ja. Du kannst Organisationen mit Mitgliedschaftsstufen erstellen und Events auf nur Mitglieder, spezifische Stufen beschränken oder sie öffentlich halten. Du hast vollständige Kontrolle über Sichtbarkeit und Zugang.'
+		},
+		{
+			question: 'Ist Revel wirklich kostenlos?',
+			answer:
+				'Ja für Selbst-Hosting (MIT-lizenziert). Unsere gehostete Version berechnet eine kleine Gebühr nur für bezahlte Ticket-Verkäufe (1,5% + 0,25€ pro Ticket). Gratis-Events und Nur-RSVP Events haben null Plattformgebühren in beiden Versionen.'
+		},
+		{
+			question: 'Welche Arten von Communities nutzen Revel?',
+			answer:
+				'Buchclubs, Laufgruppen, Makerspaces, professionelle Netzwerke, Hobby-Communities, Spezialinteressengruppen und mehr. Jede Community, die mehr als nur einfache Event-Listings will, profitiert von Revels Organisations- und Mitgliedschafts-Features.'
+		}
+	],
+	relatedPages: ['eventbrite-alternative', 'self-hosted-event-platform']
+};
+
+export const communityFirstEventPlatformIT: LandingPageContent = {
+	slug: 'community-first-event-platform',
+	locale: 'it',
+	meta: {
+		title: 'Piattaforma Eventi Community-First – Oltre il Ticketing | Revel',
+		description:
+			'Costruisci community durature, non solo eventi. Organizzazioni, livelli membership, coordinamento potluck e altro. Self-host gratis o usa la versione hosted.',
+		keywords:
+			'piattaforma eventi community, gestione membri, eventi organizzazione, coordinamento potluck, costruzione community, alternativa meetup'
+	},
+	hero: {
+		headline: 'Costruisci Community, Non Solo Eventi',
+		subheadline:
+			'Organizzazioni, membership e strumenti unici come il coordinamento potluck. Revel ti aiuta a creare connessioni community durature.'
+	},
+	intro: {
+		paragraphs: [
+			'La maggior parte delle piattaforme eventi tratta ogni incontro come transazione singola. Ma le vere community hanno bisogno di più—struttura, continuità e strumenti che supportano relazioni continuative.',
+			'Revel è costruito prima di tutto per le community. Crea organizzazioni con livelli membership, assegna ruoli e permessi, e usa funzionalità come il nostro sistema unico di coordinamento potluck per rendere gli eventi più collaborativi e meno lavoro per gli organizzatori.',
+			"Che tu gestisca un club del libro, gruppo hobby, network professionale o makerspace, Revel ti dà l'infrastruttura per crescere da meetup casuali a community fiorenti—senza abbonamenti costosi o lock-in della piattaforma."
+		]
+	},
+	features: [
+		{
+			icon: 'users',
+			title: 'Struttura Organizzazione',
+			description:
+				'Crea hub community con livelli membership, non solo listing eventi. Ruoli owner, staff e member con permessi granulari.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Coordinamento Potluck',
+			description:
+				'Sistema integrato per coordinare chi porta cosa. Gestisci restrizioni dietetiche, quantità e item—basta con spreadsheet caotici.'
+		},
+		{
+			icon: 'shield',
+			title: 'Eventi Solo-Membri',
+			description:
+				'Limita eventi a membri, livelli specifici o mantienili pubblici. Costruisci spazi esclusivi per la tua community.'
+		},
+		{
+			icon: 'ticket',
+			title: 'Ticketing Integrato',
+			description:
+				'Eventi gratis, biglietti a pagamento, solo RSVP o ibrido. Gestisci tutto dai meetup casuali alle conferenze professionali.'
+		},
+		{
+			icon: 'eye',
+			title: 'Mai Pubblicità',
+			description:
+				'La tua community merita meglio che vedere pubblicità. Self-host gratis o usa la nostra versione hosted senza pubblicità.'
+		},
+		{
+			icon: 'code',
+			title: 'Open Source (MIT)',
+			description:
+				'Gratuito da usare, modificare e deployare. Nessun vendor lock-in. Esegui sui tuoi server con controllo completo.'
+		}
+	],
+	benefits: {
+		title: 'Perché Gli Organizzatori Community Scelgono Revel',
+		items: [
+			'Costruisci strutture membership durature, non solo liste evento-per-evento',
+			'Coordina potluck e responsabilità condivise senza sforzo',
+			'Nessun abbonamento costoso come Meetup (540€/anno)',
+			'Esperienza senza pubblicità per i tuoi membri',
+			'Eventi solo-membri per costruire community esclusive',
+			'Self-host per controllo completo e zero commissioni piattaforma'
+		]
+	},
+	cta: {
+		title: 'Pronto a Costruire La Tua Community?',
+		description: 'Vedi Revel in azione o deployalo tu stesso. Nessuna carta di credito richiesta.',
+		buttons: [
+			{ text: 'Prova la Demo Live', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{ text: 'Self-Host (GitHub)', href: 'https://github.com/letsrevel', variant: 'secondary' },
+			{ text: 'Contattaci', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'Come si differenzia da Meetup?',
+			answer:
+				'Meetup costa 540€/anno e mostra pubblicità ai tuoi membri. Revel è gratuito per self-host e senza pubblicità nella nostra versione hosted. Integriamo anche funzionalità ticketing complete e feature uniche come coordinamento potluck che Meetup non ha.'
+		},
+		{
+			question: "Cos'è il coordinamento potluck?",
+			answer:
+				'È un sistema integrato che permette ai partecipanti di coordinare chi porta cosa agli eventi. Gestisci restrizioni dietetiche, gestione quantità e assegnazioni item—niente più spreadsheet caotici o strumenti esterni. È ottimo per potluck, condivisione equipment, coordinamento volontari e altro.'
+		},
+		{
+			question: 'Posso creare eventi solo-membri?',
+			answer:
+				'Sì. Puoi creare organizzazioni con livelli membership e limitare eventi a solo membri, livelli specifici o mantenerli pubblici. Hai controllo completo su visibilità e accesso.'
+		},
+		{
+			question: 'Revel è davvero gratuito?',
+			answer:
+				'Sì per self-hosting (licenza MIT). La nostra versione hosted addebita una piccola commissione solo per vendite biglietti a pagamento (1,5% + 0,25€ per biglietto). Eventi gratuiti e solo-RSVP hanno zero commissioni piattaforma in entrambe le versioni.'
+		},
+		{
+			question: 'Quali tipi di community usano Revel?',
+			answer:
+				'Club del libro, gruppi running, makerspace, network professionali, community hobby, gruppi interesse speciale e altro. Qualsiasi community che vuole più dei semplici listing eventi beneficia delle funzionalità organizzazione e membership di Revel.'
+		}
+	],
+	relatedPages: ['eventbrite-alternative', 'self-hosted-event-platform']
+};
+
+export const communityFirstEventPlatformFR: LandingPageContent = {
+	slug: 'community-first-event-platform',
+	locale: 'fr',
+	meta: {
+		title: "Plateforme d'événements axée communauté – Bien plus que la billetterie | Revel",
+		description:
+			"Crée des communautés durables, pas seulement des événements. Organisations, niveaux d'adhésion, coordination de repas partagés et bien plus. Héberge toi-même ou utilise la version hébergée.",
+		keywords:
+			"plateforme événementielle communautaire, gestion des membres, organisation d'événements, coordination repas partagé, création de communauté, alternative à meetup"
+	},
+	hero: {
+		headline: 'Crée des communautés, pas seulement des événements',
+		subheadline:
+			"Organisations, adhésions et outils uniques comme la coordination de repas partagés. Revel t'aide à tisser des liens communautaires durables."
+	},
+	intro: {
+		paragraphs: [
+			"La plupart des plateformes événementielles traitent chaque rencontre comme une transaction isolée. Mais les vraies communautés ont besoin de plus : de structure, de continuité et d'outils qui soutiennent des relations dans la durée.",
+			"Revel est pensé d'abord pour les communautés. Crée des organisations avec des niveaux d'adhésion, attribue des rôles et des permissions, et profite de fonctionnalités comme notre système unique de coordination de repas partagés pour rendre tes événements plus collaboratifs et moins de travail pour les organisateurs.",
+			"Que tu animes un club de lecture, un groupe de loisirs, un réseau professionnel ou un fablab, Revel te donne l'infrastructure pour passer de simples rencontres informelles à une communauté florissante, sans abonnements coûteux ni enfermement propriétaire."
+		]
+	},
+	features: [
+		{
+			icon: 'users',
+			title: 'Structure organisationnelle',
+			description:
+				"Crée des espaces communautaires avec des niveaux d'adhésion, pas seulement des listes d'événements. Rôles Owner, Staff et Member avec des permissions granulaires."
+		},
+		{
+			icon: 'clipboard',
+			title: 'Coordination de repas partagés',
+			description:
+				'Système intégré pour coordonner qui apporte quoi. Gère les restrictions alimentaires, les quantités et les articles : fini les tableurs chaotiques.'
+		},
+		{
+			icon: 'shield',
+			title: 'Événements réservés aux membres',
+			description:
+				'Réserve tes événements aux membres, à des niveaux précis, ou garde-les publics. Crée des espaces exclusifs pour ta communauté.'
+		},
+		{
+			icon: 'ticket',
+			title: 'Billetterie intégrée',
+			description:
+				'Événements gratuits, billets payants, RSVP seul ou formule hybride. Gère tout, des rencontres informelles aux conférences professionnelles.'
+		},
+		{
+			icon: 'eye',
+			title: 'Jamais de publicité',
+			description:
+				'Ta communauté mérite mieux que de voir des publicités. Héberge gratuitement toi-même ou utilise notre version hébergée sans publicité.'
+		},
+		{
+			icon: 'code',
+			title: 'Open source (MIT)',
+			description:
+				"Libre d'utilisation, de modification et de déploiement. Aucun enfermement propriétaire. Fais-le tourner sur tes propres serveurs avec un contrôle total."
+		}
+	],
+	benefits: {
+		title: 'Pourquoi les organisateurs de communautés choisissent Revel',
+		items: [
+			"Construis des structures d'adhésion durables, pas seulement des listes d'événements",
+			'Coordonne les repas partagés et les responsabilités collectives sans effort',
+			"Pas d'abonnements coûteux comme Meetup (540 €/an)",
+			'Une expérience sans publicité pour tes membres',
+			'Des événements réservés aux membres pour bâtir des communautés exclusives',
+			'Héberge toi-même pour un contrôle total et zéro frais de plateforme'
+		]
+	},
+	cta: {
+		title: 'Prêt à bâtir ta communauté ?',
+		description: 'Découvre Revel en action ou déploie-le toi-même. Aucune carte bancaire requise.',
+		buttons: [
+			{ text: 'Tester la démo live', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{
+				text: 'Héberger soi-même (GitHub)',
+				href: 'https://github.com/letsrevel',
+				variant: 'secondary'
+			},
+			{ text: 'Nous contacter', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'En quoi est-ce différent de Meetup ?',
+			answer:
+				"Meetup coûte 540 €/an et affiche des publicités à tes membres. Revel est gratuit en auto-hébergement et sans publicité dans sa version hébergée. Nous intégrons aussi une billetterie complète et des fonctionnalités uniques comme la coordination de repas partagés, que Meetup n'offre pas."
+		},
+		{
+			question: "Qu'est-ce que la coordination de repas partagés ?",
+			answer:
+				"C'est un système intégré qui permet aux participants de coordonner qui apporte quoi à tes événements. Gère les restrictions alimentaires, les quantités et l'attribution des articles : fini les tableurs chaotiques ou les outils externes. C'est idéal pour les repas partagés, le partage de matériel, la coordination des bénévoles et bien plus."
+		},
+		{
+			question: 'Puis-je créer des événements réservés aux membres ?',
+			answer:
+				"Oui. Tu peux créer des organisations avec des niveaux d'adhésion et réserver tes événements aux seuls membres, à des niveaux précis, ou les garder publics. Tu as un contrôle total sur la visibilité et l'accès."
+		},
+		{
+			question: 'Revel est-il vraiment gratuit ?',
+			answer:
+				"Oui pour l'auto-hébergement (licence MIT). Notre version hébergée ne facture qu'une petite commission sur les ventes de billets payants (1,5 % + 0,25 € par billet). Les événements gratuits et ceux en RSVP seul n'ont aucun frais de plateforme dans les deux versions."
+		},
+		{
+			question: 'Quels types de communautés utilisent Revel ?',
+			answer:
+				"Clubs de lecture, groupes de course, fablabs, réseaux professionnels, communautés de loisirs, groupes d'intérêt spécialisés et bien plus. Toute communauté qui veut plus que de simples listes d'événements profite des fonctionnalités d'organisation et d'adhésion de Revel."
+		}
+	],
+	relatedPages: ['eventbrite-alternative', 'self-hosted-event-platform']
+};
