+++ src/lib/data/landing-pages/queer-event-management.ts
@@ -1,0 +1,430 @@
+import type { LandingPageContent } from './types';
+
+export const queerEventManagementEN: LandingPageContent = {
+	slug: 'queer-event-management',
+	locale: 'en',
+	meta: {
+		title: 'Event Management for LGBTQ+ Communities | Revel',
+		description:
+			'Open-source event platform built by and for queer communities. Privacy controls, attendee screening, no censorship risk. Hosted in Europe.',
+		keywords:
+			'lgbtq event platform, queer event management, gay event ticketing, pride events, queer community'
+	},
+	hero: {
+		headline: 'Event Software That Gets Queer Communities',
+		subheadline:
+			"Built by LGBTQ+ organizers for events that mainstream platforms weren't designed to support."
+	},
+	intro: {
+		paragraphs: [
+			"Mainstream event platforms weren't built with queer communities in mind. Vague content policies that flag your events. Algorithms that suppress visibility. No understanding of safer spaces or community-specific needs.",
+			"Revel is different. Created by queer event organizers in Europe, it's an open-source platform designed for communities that need more than just ticketing—they need trust, privacy, and the freedom to run events without fear of deplatforming.",
+			"Whether you're organizing Pride parties, queer meetups, drag shows, or community gatherings, Revel gives you the tools to build and protect your community. Member management, attendee screening, visibility controls, and complete data ownership—all in one platform."
+		]
+	},
+	features: [
+		{
+			icon: 'heart',
+			title: 'Built for Community',
+			description:
+				'Organizations, memberships, and community-building tools. Create spaces where your community can thrive beyond single events.'
+		},
+		{
+			icon: 'shield',
+			title: 'No Censorship Risk',
+			description:
+				'Self-host or use our European servers. No corporate content policies deciding what events you can run.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Attendee Screening',
+			description:
+				'Custom questionnaires to ensure attendees align with your community values. Manual review, auto-approval, or hybrid workflows.'
+		},
+		{
+			icon: 'eye',
+			title: 'Privacy Controls',
+			description:
+				'Public, members-only, or invite-only events. Control who sees what, and keep attendee lists private.'
+		},
+		{
+			icon: 'lock',
+			title: 'Data Sovereignty',
+			description:
+				"Your community's data stays yours. No third-party trackers, no data selling, full GDPR compliance."
+		},
+		{
+			icon: 'globe',
+			title: 'European Hosting',
+			description:
+				'Hosted on European infrastructure with strong privacy protections. Self-host anywhere you choose.'
+		}
+	],
+	benefits: {
+		title: 'Why LGBTQ+ Organizers Trust Revel',
+		items: [
+			'No risk of events being flagged or removed by platform policies',
+			'Screen attendees to maintain safer spaces',
+			'Build lasting community membership, not just event-by-event lists',
+			"Full control over your community's data",
+			'Created by people who understand queer event organizing',
+			'Open source and transparent—see exactly how it works'
+		]
+	},
+	cta: {
+		title: 'Your Community Deserves Better Tools',
+		description: 'See how Revel works or start building your community today.',
+		buttons: [
+			{ text: 'Try the Live Demo', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{ text: 'Self-Host (GitHub)', href: 'https://github.com/letsrevel', variant: 'secondary' },
+			{ text: 'Contact Us', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'What makes Revel different from Eventbrite or Meetup?',
+			answer:
+				"Revel was built specifically for communities that need privacy, screening, and freedom from platform censorship. We don't have content policies that restrict adult or queer events, and we give you full control over your data."
+		},
+		{
+			question: 'Can I screen who attends my events?',
+			answer:
+				'Yes. Revel includes a powerful questionnaire system that lets you require attendees to answer questions before purchasing tickets or RSVPing. You can review submissions manually, set up auto-approval rules, or use a hybrid approach.'
+		},
+		{
+			question: "Is my community's data safe?",
+			answer:
+				"Absolutely. We don't sell data or use third-party trackers. Our hosted version runs on European servers with GDPR compliance. If you self-host, you have complete control over where your data lives."
+		},
+		{
+			question: 'Can I run members-only events?',
+			answer:
+				'Yes. You can create organizations with membership tiers and restrict events to members only, specific membership levels, or make them public. You control visibility at every level.'
+		}
+	],
+	relatedPages: ['kink-event-ticketing', 'privacy-focused-events']
+};
+
+export const queerEventManagementDE: LandingPageContent = {
+	slug: 'queer-event-management',
+	locale: 'de',
+	meta: {
+		title: 'Event-Management für LGBTQ+ Communities | Revel',
+		description:
+			'Open-Source Event-Plattform von und für queere Communities. Datenschutzkontrollen, Teilnehmer-Screening, kein Zensurrisiko. Gehostet in Europa.',
+		keywords:
+			'lgbtq event plattform, queere veranstaltungen, gay event ticketing, pride events, queer community'
+	},
+	hero: {
+		headline: 'Event-Software, Die Queere Communities Versteht',
+		subheadline:
+			'Entwickelt von LGBTQ+ Organisatoren für Events, die Mainstream-Plattformen nicht unterstützen können.'
+	},
+	intro: {
+		paragraphs: [
+			'Mainstream Event-Plattformen wurden nicht mit queeren Communities im Sinn entwickelt. Vage Inhaltsrichtlinien, die deine Events markieren. Algorithmen, die die Sichtbarkeit unterdrücken. Kein Verständnis für Safer Spaces oder community-spezifische Bedürfnisse.',
+			'Revel ist anders. Entwickelt von queeren Event-Organisatoren in Europa, ist es eine Open-Source-Plattform für Communities, die mehr als nur Ticketing brauchen – sie brauchen Vertrauen, Privatsphäre und die Freiheit, Events ohne Angst vor Deplatforming zu veranstalten.',
+			'Ob du Pride-Partys, queere Meetups, Drag-Shows oder Community-Treffen organisierst – Revel gibt dir die Werkzeuge, um deine Community aufzubauen und zu schützen. Mitgliederverwaltung, Teilnehmer-Screening, Sichtbarkeitskontrollen und vollständige Datenhoheit – alles in einer Plattform.'
+		]
+	},
+	features: [
+		{
+			icon: 'heart',
+			title: 'Für Community Gebaut',
+			description:
+				'Organisationen, Mitgliedschaften und Community-Building-Tools. Schaffe Räume, in denen deine Community über einzelne Events hinaus gedeihen kann.'
+		},
+		{
+			icon: 'shield',
+			title: 'Kein Zensurrisiko',
+			description:
+				'Selbst hosten oder unsere europäischen Server nutzen. Keine Unternehmens-Inhaltsrichtlinien, die entscheiden, welche Events du veranstalten kannst.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Teilnehmer-Screening',
+			description:
+				'Individuelle Fragebögen, um sicherzustellen, dass Teilnehmer zu den Werten deiner Community passen. Manuelle Prüfung, automatische Genehmigung oder Hybrid-Workflows.'
+		},
+		{
+			icon: 'eye',
+			title: 'Datenschutzkontrollen',
+			description:
+				'Öffentliche, nur für Mitglieder oder nur auf Einladung zugängliche Events. Kontrolliere, wer was sieht, und halte Teilnehmerlisten privat.'
+		},
+		{
+			icon: 'lock',
+			title: 'Datensouveränität',
+			description:
+				'Die Daten deiner Community bleiben deine. Keine Drittanbieter-Tracker, kein Datenverkauf, volle DSGVO-Konformität.'
+		},
+		{
+			icon: 'globe',
+			title: 'Europäisches Hosting',
+			description:
+				'Gehostet auf europäischer Infrastruktur mit starkem Datenschutz. Selbst hosten wo immer du möchtest.'
+		}
+	],
+	benefits: {
+		title: 'Warum LGBTQ+ Organisatoren Revel Vertrauen',
+		items: [
+			'Kein Risiko, dass Events durch Plattform-Richtlinien markiert oder entfernt werden',
+			'Teilnehmer screenen für sicherere Räume',
+			'Nachhaltige Community-Mitgliedschaften aufbauen, nicht nur Event-für-Event-Listen',
+			'Volle Kontrolle über die Daten deiner Community',
+			'Entwickelt von Menschen, die queere Event-Organisation verstehen',
+			'Open Source und transparent – sieh genau, wie es funktioniert'
+		]
+	},
+	cta: {
+		title: 'Deine Community Verdient Bessere Tools',
+		description: 'Sieh wie Revel funktioniert oder beginne heute mit dem Aufbau deiner Community.',
+		buttons: [
+			{ text: 'Live-Demo Testen', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{
+				text: 'Selbst Hosten (GitHub)',
+				href: 'https://github.com/letsrevel',
+				variant: 'secondary'
+			},
+			{ text: 'Kontakt', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'Was macht Revel anders als Eventbrite oder Meetup?',
+			answer:
+				'Revel wurde speziell für Communities entwickelt, die Privatsphäre, Screening und Freiheit von Plattformzensur brauchen. Wir haben keine Inhaltsrichtlinien, die Adult- oder queere Events einschränken, und wir geben dir volle Kontrolle über deine Daten.'
+		},
+		{
+			question: 'Kann ich kontrollieren, wer an meinen Events teilnimmt?',
+			answer:
+				'Ja. Revel enthält ein leistungsfähiges Fragebogensystem, mit dem du von Teilnehmern verlangen kannst, Fragen zu beantworten, bevor sie Tickets kaufen oder RSVPen. Du kannst Einreichungen manuell prüfen, Auto-Genehmigungs-Regeln einrichten oder einen Hybrid-Ansatz verwenden.'
+		},
+		{
+			question: 'Sind die Daten meiner Community sicher?',
+			answer:
+				'Absolut. Wir verkaufen keine Daten und verwenden keine Drittanbieter-Tracker. Unsere gehostete Version läuft auf europäischen Servern mit DSGVO-Konformität. Beim Selbst-Hosting hast du die komplette Kontrolle darüber, wo deine Daten liegen.'
+		},
+		{
+			question: 'Kann ich nur für Mitglieder zugängliche Events veranstalten?',
+			answer:
+				'Ja. Du kannst Organisationen mit Mitgliedschaftsstufen erstellen und Events nur auf Mitglieder, bestimmte Mitgliedschaftsstufen beschränken oder sie öffentlich machen. Du kontrollierst die Sichtbarkeit auf jeder Ebene.'
+		}
+	],
+	relatedPages: ['kink-event-ticketing', 'privacy-focused-events']
+};
+
+export const queerEventManagementIT: LandingPageContent = {
+	slug: 'queer-event-management',
+	locale: 'it',
+	meta: {
+		title: 'Gestione Eventi per Community LGBTQ+ | Revel',
+		description:
+			'Piattaforma eventi open-source creata da e per community queer. Controlli privacy, screening partecipanti, nessun rischio censura. Hosting in Europa.',
+		keywords:
+			'piattaforma eventi lgbtq, gestione eventi queer, ticketing eventi gay, eventi pride, community queer'
+	},
+	hero: {
+		headline: 'Software Eventi Che Capisce le Community Queer',
+		subheadline:
+			'Creato da organizzatori LGBTQ+ per eventi che le piattaforme mainstream non sanno supportare.'
+	},
+	intro: {
+		paragraphs: [
+			'Le piattaforme eventi mainstream non sono state create pensando alle community queer. Policy sui contenuti vaghe che segnalano i tuoi eventi. Algoritmi che sopprimono la visibilità. Nessuna comprensione degli spazi sicuri o delle esigenze specifiche della community.',
+			'Revel è diverso. Creato da organizzatori di eventi queer in Europa, è una piattaforma open-source progettata per community che hanno bisogno di più del semplice ticketing—hanno bisogno di fiducia, privacy e libertà di organizzare eventi senza paura di essere rimossi.',
+			"Che tu stia organizzando feste Pride, meetup queer, drag show o ritrovi della community, Revel ti dà gli strumenti per costruire e proteggere la tua community. Gestione membri, screening partecipanti, controlli visibilità e proprietà completa dei dati—tutto in un'unica piattaforma."
+		]
+	},
+	features: [
+		{
+			icon: 'heart',
+			title: 'Costruito per la Community',
+			description:
+				'Organizzazioni, membership e strumenti di community building. Crea spazi dove la tua community può prosperare oltre i singoli eventi.'
+		},
+		{
+			icon: 'shield',
+			title: 'Nessun Rischio Censura',
+			description:
+				'Self-host o usa i nostri server europei. Nessuna policy aziendale che decide quali eventi puoi organizzare.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Screening Partecipanti',
+			description:
+				'Questionari personalizzati per assicurare che i partecipanti siano in linea con i valori della community. Revisione manuale, auto-approvazione o workflow ibridi.'
+		},
+		{
+			icon: 'eye',
+			title: 'Controlli Privacy',
+			description:
+				'Eventi pubblici, solo membri o solo su invito. Controlla chi vede cosa e mantieni private le liste partecipanti.'
+		},
+		{
+			icon: 'lock',
+			title: 'Sovranità dei Dati',
+			description:
+				'I dati della tua community restano tuoi. Nessun tracker di terze parti, nessuna vendita dati, piena conformità GDPR.'
+		},
+		{
+			icon: 'globe',
+			title: 'Hosting Europeo',
+			description:
+				'Hostato su infrastruttura europea con forti protezioni privacy. Self-host ovunque tu scelga.'
+		}
+	],
+	benefits: {
+		title: 'Perché gli Organizzatori LGBTQ+ Si Fidano di Revel',
+		items: [
+			'Nessun rischio che gli eventi vengano segnalati o rimossi per policy della piattaforma',
+			'Screening partecipanti per mantenere spazi più sicuri',
+			'Costruisci membership durature, non solo liste evento per evento',
+			'Controllo totale sui dati della tua community',
+			"Creato da persone che capiscono l'organizzazione eventi queer",
+			'Open source e trasparente—vedi esattamente come funziona'
+		]
+	},
+	cta: {
+		title: 'La Tua Community Merita Strumenti Migliori',
+		description: 'Scopri come funziona Revel o inizia a costruire la tua community oggi.',
+		buttons: [
+			{ text: 'Prova la Demo', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{ text: 'Self-Host (GitHub)', href: 'https://github.com/letsrevel', variant: 'secondary' },
+			{ text: 'Contattaci', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'Cosa rende Revel diverso da Eventbrite o Meetup?',
+			answer:
+				'Revel è stato costruito specificamente per community che necessitano di privacy, screening e libertà dalla censura delle piattaforme. Non abbiamo policy sui contenuti che limitano eventi adult o queer, e ti diamo il controllo totale sui tuoi dati.'
+		},
+		{
+			question: 'Posso controllare chi partecipa ai miei eventi?',
+			answer:
+				'Sì. Revel include un potente sistema di questionari che ti permette di richiedere ai partecipanti di rispondere a domande prima di acquistare biglietti o fare RSVP. Puoi revisionare le risposte manualmente, impostare regole di auto-approvazione o usare un approccio ibrido.'
+		},
+		{
+			question: 'I dati della mia community sono al sicuro?',
+			answer:
+				'Assolutamente. Non vendiamo dati né usiamo tracker di terze parti. La nostra versione hosted gira su server europei con conformità GDPR. Se fai self-host, hai il controllo completo su dove risiedono i tuoi dati.'
+		},
+		{
+			question: 'Posso organizzare eventi solo per membri?',
+			answer:
+				'Sì. Puoi creare organizzazioni con livelli di membership e limitare gli eventi ai soli membri, a livelli specifici di membership, o renderli pubblici. Controlli la visibilità a ogni livello.'
+		}
+	],
+	relatedPages: ['kink-event-ticketing', 'privacy-focused-events']
+};
+
+export const queerEventManagementFR: LandingPageContent = {
+	slug: 'queer-event-management',
+	locale: 'fr',
+	meta: {
+		title: 'Gestion d’événements pour les communautés LGBTQ+ | Revel',
+		description:
+			'Plateforme d’événements open source par et pour les communautés queer. Contrôles de confidentialité, sélection des participant·es, aucun risque de censure. Hébergée en Europe.',
+		keywords:
+			'plateforme événements lgbtq, événements queer, billetterie événements gay, événements pride, communauté queer'
+	},
+	hero: {
+		headline: 'Un logiciel d’événements qui comprend les communautés queer',
+		subheadline:
+			'Conçu par des organisateur·ices LGBTQ+ pour des événements que les plateformes grand public ne peuvent pas accueillir.'
+	},
+	intro: {
+		paragraphs: [
+			'Les plateformes d’événements grand public n’ont pas été pensées pour les communautés queer. Des règles de contenu vagues qui signalent tes événements. Des algorithmes qui étouffent ta visibilité. Aucune compréhension des espaces plus sûrs ni des besoins propres à chaque communauté.',
+			'Revel, c’est différent. Conçue par des organisateur·ices d’événements queer en Europe, c’est une plateforme open source pour les communautés qui ont besoin de plus qu’une simple billetterie – elles ont besoin de confiance, de respect de la vie privée et de la liberté d’organiser des événements sans craindre d’être déplateformées.',
+			'Que tu organises des soirées Pride, des rencontres queer, des spectacles de drag ou des réunions de communauté, Revel te donne les outils pour bâtir et protéger ta communauté. Gestion des membres, sélection des participant·es, contrôles de visibilité et pleine souveraineté sur tes données – le tout sur une seule plateforme.'
+		]
+	},
+	features: [
+		{
+			icon: 'heart',
+			title: 'Pensé pour la communauté',
+			description:
+				'Organisations, adhésions et outils de développement communautaire. Crée des espaces où ta communauté peut s’épanouir au-delà des événements ponctuels.'
+		},
+		{
+			icon: 'shield',
+			title: 'Aucun risque de censure',
+			description:
+				'Héberge toi-même ou utilise nos serveurs européens. Aucune règle de contenu d’entreprise ne décide quels événements tu as le droit d’organiser.'
+		},
+		{
+			icon: 'clipboard',
+			title: 'Sélection des participant·es',
+			description:
+				'Des questionnaires personnalisés pour t’assurer que les participant·es correspondent aux valeurs de ta communauté. Validation manuelle, approbation automatique ou flux hybrides.'
+		},
+		{
+			icon: 'eye',
+			title: 'Contrôles de confidentialité',
+			description:
+				'Événements publics, réservés aux membres ou sur invitation uniquement. Choisis qui voit quoi et garde tes listes de participant·es privées.'
+		},
+		{
+			icon: 'lock',
+			title: 'Souveraineté des données',
+			description:
+				'Les données de ta communauté restent les tiennes. Aucun traceur tiers, aucune revente de données, conformité totale au RGPD.'
+		},
+		{
+			icon: 'globe',
+			title: 'Hébergement européen',
+			description:
+				'Hébergé sur une infrastructure européenne avec une protection des données solide. Ou héberge toi-même où tu le souhaites.'
+		}
+	],
+	benefits: {
+		title: 'Pourquoi les organisateur·ices LGBTQ+ font confiance à Revel',
+		items: [
+			'Aucun risque que tes événements soient signalés ou supprimés par les règles d’une plateforme',
+			'Sélectionner les participant·es pour des espaces plus sûrs',
+			'Bâtir des adhésions communautaires durables, pas seulement des listes événement par événement',
+			'Garder le contrôle total sur les données de ta communauté',
+			'Conçu par des personnes qui comprennent l’organisation d’événements queer',
+			'Open source et transparent – vois exactement comment ça fonctionne'
+		]
+	},
+	cta: {
+		title: 'Ta communauté mérite de meilleurs outils',
+		description:
+			'Découvre comment fonctionne Revel ou commence dès aujourd’hui à bâtir ta communauté.',
+		buttons: [
+			{ text: 'Tester la démo en direct', href: 'https://demo.letsrevel.io', variant: 'primary' },
+			{
+				text: 'Héberger soi-même (GitHub)',
+				href: 'https://github.com/letsrevel',
+				variant: 'secondary'
+			},
+			{ text: 'Nous contacter', href: 'mailto:contact@letsrevel.io', variant: 'outline' }
+		]
+	},
+	faq: [
+		{
+			question: 'Qu’est-ce qui distingue Revel d’Eventbrite ou de Meetup ?',
+			answer:
+				'Revel a été conçu spécialement pour les communautés qui ont besoin de confidentialité, de sélection des participant·es et d’une liberté face à la censure des plateformes. Nous n’avons aucune règle de contenu qui limite les événements pour adultes ou queer, et nous te donnons le contrôle total de tes données.'
+		},
+		{
+			question: 'Puis-je contrôler qui participe à mes événements ?',
+			answer:
+				'Oui. Revel intègre un puissant système de questionnaires qui te permet d’exiger des participant·es qu’iels répondent à des questions avant d’acheter un billet ou de confirmer leur présence. Tu peux examiner les réponses manuellement, définir des règles d’approbation automatique ou opter pour une approche hybride.'
+		},
+		{
+			question: 'Les données de ma communauté sont-elles en sécurité ?',
+			answer:
+				'Absolument. Nous ne vendons aucune donnée et n’utilisons aucun traceur tiers. Notre version hébergée tourne sur des serveurs européens conformes au RGPD. En auto-hébergement, tu gardes le contrôle total de l’emplacement de tes données.'
+		},
+		{
+			question: 'Puis-je organiser des événements réservés aux membres ?',
+			answer:
+				'Oui. Tu peux créer des organisations avec des niveaux d’adhésion et réserver des événements aux membres, à certains niveaux d’adhésion précis, ou les rendre publics. Tu contrôles la visibilité à chaque niveau.'
+		}
+	],
+	relatedPages: ['kink-event-ticketing', 'privacy-focused-events']
+};
